CLOUDIFY-852
Cache USM monitors result (expiration defaults to 5 seconds)

git-svn-id: svn://pc-lab14/SVN/cloudify/trunk/cloudify@122647 eb64e737-3616-4df0-8941-5ee2ae88103d

+++ USM/src/main/java/org/cloudifysource/usm/dsl/DSLConfiguration.java
@@ -33,7 +33,7 @@ import org.cloudifysource.usm.UniversalServiceManagerConfiguration;
  */
 public class DSLConfiguration implements UniversalServiceManagerConfiguration {
 
-	private static final int MILLIS_IN_SECOND = 1000;
+	private static final long MILLIS_IN_SECOND = 1000L;
 	private final Service service;
 	private final File puExtDir;
 	private final ServiceContextImpl serviceContext;
